requirements.txt
requests
discord.py[speed]
aiohttp
aiosqlite
aiofiles
pyyaml
beautifulsoup4